Struct node_rs::chain::chain::Chain

pub struct Chain {
    pub genesis_configuration_hash: String,
    pub genesis_identifier_hash: String,
    pub blocks: HashMap<String, Block>,
    pub adjacent_matrix: HashMap<String, Vec<String>>,
}

Fields
genesis_configuration_hash: String

the hash of the genesis configuration

genesis_identifier_hash: String

the hash of the genesis block

blocks: HashMap<String, Block>

all known blocks

adjacent_matrix: HashMap<String, Vec<String>>

a matrix creating the relation between blocks key is the parent, values are its children

Methods

impl Chain

pub fn new(genesis_hash: String) -> Self

pub fn get_current_block_number(&self) -> usize

pub fn get_current_block_timestamp(&self) -> u64

pub fn get_current_block(&self) -> (usize, Block)

pub fn has_parent_of_block(self, block: Block) -> bool

Returns true, if the parent of the given block exists, false otherwise.

pub fn add_block(&mut self, block: Block) -> bool

Add the block as child to its corresponding parent. Panics, if the parent block specified does not exist. Therefore, invoke has_parent_of_block first.

Returns true, if the block was added, false otherwise.