Module node_rs::chain::chain_visitor

Visitors of the chain which can be used in combination with a chain walker.

Structs

HeaviestBlockVisitor
This visitor expects to be called exactly once with the heaviest block in the chain.

SumCipherTextVisitor
Sums up all votes contained in the transactions, after the voting has been opened and until it is closed again.

Traits

ChainVisitor