Struct node_rs::chain::chain_visitor::HeaviestBlockVisitor

pub struct HeaviestBlockVisitor {
    pub height: Option<usize>,
    pub heaviest_block: Option<String>,
}

This visitor expects to be called exactly once with the heaviest block in the chain.

After it is called, it's heaviest_block contains a hash of the heaviest block currently in the chain.

Fields
height: Option<usize>

The height of the heaviest block.

heaviest_block: Option<String>

The hash of the string once it is assigned, or None, if this visitor was never visited.

Methods

impl HeaviestBlockVisitor

pub fn new() -> HeaviestBlockVisitor

Create a new HeaviestBlockVisitor having a None hash of the heaviest block.