Struct node_rs::p2p::node::Node

pub struct Node { /* fields omitted */ }

Forms a node in the blockchain.

Each node manages its own thread pool on which it starts dedicated threads to listen for incoming connections. In addition, connection attempts to other nodes are also spawn on the thread pool.

Methods

impl Node

pub fn new(
listen_address: SocketAddr,
rpc_listen_address: SocketAddr,
genesis: Genesis
) -> Node

Creates a new node.

listen_addr The address on which the node listens for incoming messages.

rpc_listen_address The address on which the node listens for incoming RPC messages.

genesis The genesis configuration which defines the behaviour of this node. Must be equal for all nodes which should connect to the same network.

pub fn listen(&self)

Start a listener on the bootstrap address.

Read all bytes until EOF (when underlying socket is closed) from the given stream and return a message back to the incoming sender. Then close the stream in order to signal EOF for the receiving node.

pub fn listen_rpc(&self)

Start to listen for incoming RPC connections, i.e. connections from an end-user client. Compared to pub fn listen(&self), incoming messages may be handled a bit differently.

pub fn request_chain_copy(&mut self)

Send a request for a copy of the blockchain to all known nodes.

pub fn sign(&mut self)

Start the main loop to sign (aka. mint) blocks in the network.