Struct node_rs::protocol::clique::CliqueProtocol

pub struct CliqueProtocol { /* fields omitted */ }

The clique protocol provides a Proof-of-Authority (PoA) sybil control mechanism.

Methods

impl CliqueProtocol

pub fn new(own_address: SocketAddr, genesis: Genesis) -> Self

Create a new protocol instance.

own_address: The socket address the node is listening on. Used to find the own sealer index in the genesis configuration.

genesis: The initial configuration of the clique protocol.

Panics

Panics if the given own_address is not contained in the genesis configuration.

pub fn replace_chain(&mut self, chain: Chain)

Replace the own block chain with the given instance, if the given instance has a branch with a greater height than our longest branch.

pub fn is_leader(&self) -> bool

Returns true, if the node is a leader in the current epoch and therefore allowed to sign blocks.

pub fn is_co_leader(&self) -> bool

Returns true, if the node is a co-leader in the current epoch and therefore allowed to sign a blocks after waiting for a particular wiggle time.

pub fn is_block_period_over(&self) -> bool

pub fn create_current_block_and_reset_transaction_buffer(&mut self) -> Block

pub fn reset_transaction_buffer(&mut self)

pub fn sign(&mut self, block: Block) -> Option<Block>

Sign a block with all current known transactions. May return None if a block with the same identifier is already contained in the chain of the node.